koordinat/requirements.txt

djangorestframework
djangorestframework-simplejwt   